Windows: use the newer style of port path name for serial port
the common Windows device name definition is "\\\\.\\COMx" (without trailing colon) for a serial port
This removes the limitation of port names from COM1 to COM9, because a PC has a lot of bluetooth ports in use too

The code set the correct new path name and strips the possible trailing colon,
if there is an old setting in the profile file.
With a further commit the setting name will not have a trailing colon.