Use of the user struct also for the setting of the pilot.
Modification of all affected files.
------------------------------------------------------------------------

W1: net/client/WeGlide/WeGlideSettings create a WeGlide User struct,  Aircraft struct and a Flight struct similar to the WeGlide objects (line too long!)

this structures should be a representation of the User, Aircraft and Flight structure at the WeGlide server with similar fields. This can also be used here too instead of creation similar fields doubled

This structures can be used for Upload and Response and
for saving to a User list (later...) for a fast access to
all used users (pilots)

use of predefined values